정령 난수 방식변경 전투시 난수 생성 방식 변경

+++ UnlimitedTowerContract/Contract_eosiocdt/mainnet_ver1.0/battletest.hpp
@@ -1368,7 +1368,7 @@ CONTRACT battletest : public contract
         uint32_t avoid = 0;
         uint32_t speed = 0;
         uint32_t position = 0;
-        uint32_t second_speed = 0;
+        uint64_t second_speed = 0;
         uint32_t action = 0;
         status_info status;
         uint32_t id = 0;
